+++ javascript/src/types.ts
@@ -9,3 +9,18 @@ export interface Expression {
   readonly source: string
   match(text: string): readonly Argument[] | null
 }
+
+export type ParameterInfo = {
+  /**
+   * The string representation of the original ParameterType#type property
+   */
+  type: string | null
+  /**
+   * The parameter type name
+   */
+  name: string
+  /**
+   * The number of times this name has been used so far
+   */
+  count: number
+}